feat: Remove all non-native big-integer fallbacks
